LayoutTests/imported/w3c/web-platform-tests/eventsource/request-credentials-expected.txt

Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=1_1505770315694
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=1_1505770315694.
Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=2_1505770315695
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=2_1505770315695.
Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=3_1505770315695
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=3_1505770315695.

Harness Error (TIMEOUT), message = null

TIMEOUT EventSource: credentials: credentials enabled Test timed out
TIMEOUT EventSource: credentials: credentials disabled Test timed out
TIMEOUT EventSource: credentials: credentials default Test timed out
